ucache: division by MultiCache/Cache
