+++ ucache/ucache.go
@@ -1,9 +1,13 @@
+/*
+ * @kordax (Dmitry Morozov)
+ * dmorozov@valoru-software.com
+ * Copyright (c) 2024.
+ */
+
 package ucache
 
 import (
 	"bytes"
-	"crypto/sha256"
-	"encoding/binary"
 	"sync"
 	"time"
 
@@ -12,324 +16,43 @@ import (
 	"github.com/kordax/basic-utils/uopt"
 )
 
-type container[K CompositeKey, T Comparable] struct {
-	pairs map[int64][]uarray.Pair[K, T]
-	node  map[int64]any
-}
+// The Cache interface defines a set of methods for a generic cache implementation.
+// This interface supports setting, getting, and managing cache entries with composite keys.
+// Unlike MultiCache, it is designed to handle only one value per key.
+type Cache[K CompositeKey, T any] interface {
+	// Set updates the cache value for the provided key. If the key already exists,
+	// its previous value is removed before adding the new value. This method should be thread-safe.
+	Set(key K, value T)
 
-type Cache[K CompositeKey, T Comparable] interface {
-	Put(key K, values ...T) // Adds a new value to the key
-	Set(key K, values ...T) // Overwrites key values
-	Get(key K) []T
-	Changes() []K
+	// Get retrieves the value associated with the provided key from the cache.
+	// It returns the value and a boolean indicating whether the key was found.
+	// This method should be thread-safe.
+	Get(key K) (*T, bool)
+
+	// Drop completely clears the cache, removing all entries. This method should be thread-safe.
 	Drop()
+
+	// DropKey removes the value associated with the provided key from the cache. This method should be thread-safe.
 	DropKey(key K)
+
+	// Outdated checks if the provided key or the entire cache (if no key is provided)
+	// is outdated based on the set TTL (time-to-live). Returns true if outdated, false otherwise.
+	// This method should be thread-safe.
 	Outdated(key uopt.Opt[K]) bool
 
-	PutQuietly(key K, values ...T) // Adds the values but update cache state/doesn't add any changes to the cache
+	// SetQuietly is an optimized method adds a value to the cache for the provided key but does so without
+	// altering the change history. This method is useful when modifications should not trigger cache change diff.
+	// This method should be thread-safe.
+	// This operation is much faster and can be used to optimize cache performance in case you don't want to track changes.
+	SetQuietly(key K, value T)
 }
 
-// InMemoryTreeCache provides an in-memory caching mechanism with support for compound keys.
-// The cache leverages tree-like structures to store and organize data, allowing efficient
-// operations even with composite keys. The cache supports optional TTL (time-to-live) for entries,
-// ensuring that outdated entries can be identified and potentially purged. Concurrency-safe
-// operations are ensured through the use of a mutex.
-//
-// Benchmark insights:
-// - Put operation performance is fast for shallow depth keys but slows down as the depth increases.
-// - Get operation is particularly efficient, especially for shallow depth keys.
-// - Set operation's performance is consistent regardless of the depth of the key.
-type InMemoryTreeCache[K CompositeKey, T Comparable] struct {
-	values  map[int64]any
-	changes []K
-
-	lastUpdatedKeys map[string]time.Time
-	lastUpdated     time.Time
-	ttl             *time.Duration
-
-	vMtx sync.Mutex
-}
-
-// NewInMemoryTreeCache creates a new instance of the InMemoryTreeCache.
-// It takes an optional TTL (time-to-live) parameter to set expiration time for cache entries.
-// If the TTL is not provided, cache entries will not expire.
-func NewInMemoryTreeCache[K CompositeKey, T Comparable](ttl uopt.Opt[time.Duration]) *InMemoryTreeCache[K, T] {
-	c := &InMemoryTreeCache[K, T]{
-		values:          make(map[int64]any),
-		changes:         make([]K, 0),
-		lastUpdatedKeys: make(map[string]time.Time),
-	}
-	ttl.IfPresent(func(t time.Duration) {
-		c.ttl = &t
-	})
-
-	return c
-}
-
-// Put inserts a new value(s) into the cache associated with the given key.
-// If the key already exists in the cache, it appends the new value(s) to the existing values.
-// This operation is relatively fast for shallow depth keys, but becomes slower as the depth increases.
-func (c *InMemoryTreeCache[K, T]) Put(key K, val ...T) {
-	c.vMtx.Lock()
-	defer c.vMtx.Unlock()
-	c.put(key, val...)
-	c.lastUpdatedKeys[key.String()] = time.Now()
-	c.lastUpdated = time.Now()
-}
-
-// Set inserts a new value(s) into the cache associated with the given key.
-// If the key already exists in the cache, this method will overwrite the existing values.
-func (c *InMemoryTreeCache[K, T]) Set(key K, val ...T) {
-	c.vMtx.Lock()
-	defer c.vMtx.Unlock()
-	c.dropKeyRecursively(key.Keys(), 0, c.values)
-	c.put(key, val...)
-	c.lastUpdatedKeys[key.String()] = time.Now()
-	c.lastUpdated = time.Now()
-}
-
-// AddSilently behaves like the Put method but does not update the cache state or add any changes to the cache.
-// This method is useful when you want to add values to the cache without triggering any side effects.
-func (c *InMemoryTreeCache[K, T]) PutQuietly(key K, val ...T) {
-	c.vMtx.Lock()
-	defer c.vMtx.Unlock()
-	c.addTran(key, val...)
-	c.lastUpdatedKeys[key.String()] = time.Now()
-	c.lastUpdated = time.Now()
-}
-
-// Get retrieves the value(s) associated with the given key from the cache.
-// If the key is not found, it returns an empty slice.
-// Retrieval is fast, especially for shallow depth keys.
-func (c *InMemoryTreeCache[K, T]) Get(key K) []T {
-	c.vMtx.Lock()
-	defer c.vMtx.Unlock()
-	c.changes = nil
-	bucket := c.tryToGetBucket(key.Keys())
-	result := make([]T, 0)
-	for _, pairs := range bucket {
-		for _, p := range pairs {
-			result = append(result, p.Right)
-		}
-	}
-
-	return result
-}
-
-// Changes returns a slice of keys that have been modified in the cache.
-// This method provides a way to track changes made to the cache, useful for scenarios like cache syncing.
-func (c *InMemoryTreeCache[K, T]) Changes() []K {
-	return c.changes
-}
-
-// Drop removes all entries from the cache.
-// This is a complete reset of the cache, useful when you want to clear the cache and start fresh.
-func (c *InMemoryTreeCache[K, T]) Drop() {
-	c.vMtx.Lock()
-	defer c.vMtx.Unlock()
-	c.dropAll()
-	c.lastUpdatedKeys = make(map[string]time.Time)
-}
-
-// DropKey removes the value(s) associated with the given key from the cache.
-func (c *InMemoryTreeCache[K, T]) DropKey(key K) {
-	c.vMtx.Lock()
-	defer c.vMtx.Unlock()
-	c.dropKeyRecursively(key.Keys(), 0, c.values)
-	c.lastUpdatedKeys[key.String()] = time.Now()
-	c.lastUpdated = time.Now()
-}
-
-// Outdated checks if a given key or the entire cache is outdated based on the TTL.
-// If no key is provided or key was not found, it checks the last updated time of the entire cache.
-// If a key is provided and found, it checks the last updated time of that specific key.
-func (c *InMemoryTreeCache[K, T]) Outdated(key uopt.Opt[K]) bool {
-	if !key.Present() {
-		return time.Since(c.lastUpdated) > *c.ttl
-	}
-
-	c.vMtx.Lock()
-	defer c.vMtx.Unlock()
-
-	if c.ttl == nil {
-		return false
-	} else {
-		if key.Present() {
-			k := key.Get()
-			if lu, ok := c.lastUpdatedKeys[(*k).String()]; ok {
-				return time.Since(lu) > *c.ttl
-			} else {
-				return true
-			}
-		} else {
-			return time.Since(c.lastUpdated) > *c.ttl
-		}
-	}
-}
-
-func (c *InMemoryTreeCache[K, T]) dropAll() {
-	c.values = make(map[int64]any)
-	c.changes = nil
-}
-
-func (c *InMemoryTreeCache[K, T]) put(key K, val ...T) {
-	c.addTran(key, val...)
-	changes := len(c.changes) == 0
-	found := false
-	for _, diff := range c.changes {
-		if uarray.EqualsWithOrder(diff.Keys(), key.Keys()) {
-			if !diff.Equals(key) {
-				changes = true
-				break
-			}
-			found = true
-			continue
-		}
-	}
-	if changes || !found {
-		c.changes = append(c.changes, key)
-	}
-}
-
-func (c *InMemoryTreeCache[K, T]) addTran(key K, values ...T) {
-	hashes := key.Keys()
-	if len(hashes) == 0 {
-		return
-	}
-
-	bucket := c.tryToGetBucket(hashes)
-	lHash := key.Keys()[len(hashes)-1]
-
-	for _, value := range values {
-		if ind, _ := uarray.ContainsPredicate(bucket[lHash], func(v *uarray.Pair[K, T]) bool {
-			return v.Right.Equals(value)
-		}); ind > -1 {
-			bucket[lHash][ind] = *uarray.NewPair[K, T](key, value)
-		} else {
-			bucket[lHash] = append(bucket[lHash], *uarray.NewPair[K, T](key, value))
-		}
-	}
-}
-
-func (c *InMemoryTreeCache[K, T]) dropKeyRecursively(keys []int64, n int, bucket map[int64]any) {
-	hash := keys[n]
-	interBucket := bucket[hash]
-	if interBucket != nil {
-		switch b := interBucket.(type) {
-		case container[K, T]:
-			if n+1 == len(keys) {
-				delete(bucket, hash)
-			} else {
-				c.dropKeyRecursively(keys, n+1, b.node)
-			}
-		default:
-			delete(bucket, hash)
-		}
-	}
-}
-
-func (c *InMemoryTreeCache[K, T]) tryToGetBucket(keys []int64) map[int64][]uarray.Pair[K, T] {
-	return c.getBucket(keys, 0, c.values)
-}
-
-func (c *InMemoryTreeCache[K, T]) getBucket(keys []int64, n int, interBucket map[int64]any) map[int64][]uarray.Pair[K, T] {
-	if keys == nil || n >= len(keys) {
-		return nil
-	}
-
-	if bucket, ok := interBucket[keys[n]]; ok {
-		switch b := bucket.(type) {
-		case map[int64][]uarray.Pair[K, T]:
-			if n+1 < len(keys) {
-				interBucket[keys[n]] = container[K, T]{
-					node:  make(map[int64]any),
-					pairs: b,
-				}
-				return c.getBucket(keys, n+1, interBucket[keys[n]].(container[K, T]).node)
-			} else {
-				return b
-			}
-		case container[K, T]:
-			if n+1 == len(keys) {
-				result := make(map[int64][]uarray.Pair[K, T])
-				for k, e := range b.pairs {
-					result[k] = append(result[k], e...)
-				}
-				if b.node != nil {
-					result = c.getNodePairsFlat(b.node, result)
-				}
-
-				return result
-			}
-
-			return c.getBucket(keys, n+1, b.node)
-		}
-	} else {
-		if n+1 == len(keys) {
-			interBucket[keys[n]] = map[int64][]uarray.Pair[K, T]{
-				keys[n]: nil,
-			}
-			return interBucket[keys[n]].(map[int64][]uarray.Pair[K, T])
-		} else {
-			if entry, ok := interBucket[keys[n]]; !ok {
-				interBucket[keys[n]] = container[K, T]{
-					node:  make(map[int64]any),
-					pairs: make(map[int64][]uarray.Pair[K, T]),
-				}
-				return c.getBucket(keys, n+1, interBucket[keys[n]].(container[K, T]).node)
-			} else {
-				switch e := entry.(type) {
-				case map[int64][]uarray.Pair[K, T]:
-					interBucket[keys[n]] = container[K, T]{
-						node:  make(map[int64]any),
-						pairs: e,
-					}
-					return c.getBucket(keys, n+1, interBucket[keys[n]].(container[K, T]).node)
-				case container[K, T]:
-					interBucket[keys[n]] = container[K, T]{
-						pairs: e.pairs,
-					}
-					return c.getBucket(keys, n+1, e.node)
-				}
-			}
-		}
-	}
-
-	return nil
-}
-
-func (c *InMemoryTreeCache[K, T]) getNodePairsFlat(node map[int64]any, result map[int64][]uarray.Pair[K, T]) map[int64][]uarray.Pair[K, T] {
-	for _, entry := range node {
-		switch e := entry.(type) {
-		case map[int64][]uarray.Pair[K, T]:
-			for hash, pair := range e {
-				result[hash] = append(result[hash], pair...)
-			}
-		case container[K, T]:
-			for hash, pair := range e.pairs {
-				result[hash] = append(result[hash], pair...)
-			}
-			result = c.getNodePairsFlat(e.node, result)
-		}
-	}
-
-	return result
-}
-
-// InMemoryHashMapCache provides an in-memory caching mechanism using hashmaps.
-// This cache structure translates composite keys into a hash value using a user-provided
-// hashing function. The cache supports optional TTL (time-to-live) for entries.
-// Concurrency-safe operations are ensured through the use of a mutex.
-//
-// Performance Comparison with InMemoryTreeCache:
-// - Insertions: InMemoryTreeCache is slightly faster for single-depth insertions and significantly faster for deeper depths.
-// - Retrievals: InMemoryHashMapCache (especially with FarmHash) is faster for both single-depth and deeper retrievals.
-// - Setting Values: Performance is relatively close between the two, with minor variations.
-//
-// The choice between InMemoryTreeCache and InMemoryHashMapCache would depend on the specific use case,
-// especially the depth of the keys and the frequency of retrieval operations.
-type InMemoryHashMapCache[K CompositeKey, T Comparable, H comparable] struct {
-	values  map[H][]T
+// InMemoryHashMapCache provides an in-memory caching mechanism using hashmaps for single-value entries.
+// Unlike InMemoryHashMapMultiCache, it stores only one value per key.
+// This structure translates composite keys into a hash value using a user-provided hashing function.
+// Supports optional TTL for entries and ensures concurrency-safe operations using a mutex.
+type InMemoryHashMapCache[K CompositeKey, T any, H comparable] struct {
+	values  map[H]T
 	changes []K
 
 	lastUpdatedKeys map[string]time.Time
@@ -343,9 +66,9 @@ type InMemoryHashMapCache[K CompositeKey, T Comparable, H comparable] struct {
 // NewInMemoryHashMapCache creates a new instance of the InMemoryHashMapCache.
 // It takes a hashing function to translate the composite keys to a desired hash type,
 // and an optional time-to-live duration for the cache entries.
-func NewInMemoryHashMapCache[K CompositeKey, T Comparable, H comparable](toHash func(keys []int64) H, ttl uopt.Opt[time.Duration]) *InMemoryHashMapCache[K, T, H] {
+func NewInMemoryHashMapCache[K CompositeKey, T any, H comparable](toHash func(keys []int64) H, ttl uopt.Opt[time.Duration]) Cache[K, T] {
 	c := &InMemoryHashMapCache[K, T, H]{
-		values:          make(map[H][]T),
+		values:          make(map[H]T),
 		changes:         make([]K, 0),
 		lastUpdatedKeys: make(map[string]time.Time),
 		toHash:          toHash,
@@ -358,11 +81,11 @@ func NewInMemoryHashMapCache[K CompositeKey, T Comparable, H comparable](toHash 
 }
 
 // NewDefaultHashMapCache creates a new instance of the InMemoryHashMapCache using SHA256 as the hashing algorithm.
-func NewDefaultHashMapCache[K CompositeKey, T Comparable](ttl uopt.Opt[time.Duration]) *InMemoryHashMapCache[K, T, uint64] {
+func NewDefaultHashMapCache[K CompositeKey, T any](ttl uopt.Opt[time.Duration]) Cache[K, T] {
 	return NewFarmHashMapCache[K, T](ttl)
 }
 
-func NewFarmHashMapCache[K CompositeKey, T Comparable](ttl uopt.Opt[time.Duration]) *InMemoryHashMapCache[K, T, uint64] {
+func NewFarmHashMapCache[K CompositeKey, T any](ttl uopt.Opt[time.Duration]) Cache[K, T] {
 	buffer := new(bytes.Buffer)
 	return NewInMemoryHashMapCache[K, T, uint64](func(keys []int64) uint64 {
 		arr := make([]byte, 0)
@@ -374,68 +97,36 @@ func NewFarmHashMapCache[K CompositeKey, T Comparable](ttl uopt.Opt[time.Duratio
 	}, ttl)
 }
 
-func NewSha256HashMapCache[K CompositeKey, T Comparable](ttl uopt.Opt[time.Duration]) *InMemoryHashMapCache[K, T, string] {
-	buffer := new(bytes.Buffer)
-	return NewInMemoryHashMapCache[K, T, string](func(keys []int64) string {
-		arr := make([]byte, 0)
-		for _, hash := range keys {
-			arr = append(arr, intToBytes(buffer, hash)...)
-		}
-
-		h := sha256.New()
-		h.Write(arr)
-
-		return string(h.Sum(nil))
-	}, ttl)
-}
-
-// Put adds the given values to the cache associated with the provided key.
-// If the key already exists, the values are updated. The insertion is thread-safe.
-func (c *InMemoryHashMapCache[K, T, H]) Put(key K, values ...T) {
-	if len(values) == 0 {
-		return
-	}
+// Set updates the cache value for the provided key. If the key already exists,
+// its previous value are removed before adding the new value. The operation is thread-safe.
+func (c *InMemoryHashMapCache[K, T, H]) Set(key K, value T) {
 	c.vMtx.Lock()
 	defer c.vMtx.Unlock()
-	c.put(key, values...)
+	c.put(key, value)
 	c.lastUpdatedKeys[key.String()] = time.Now()
 	c.lastUpdated = time.Now()
 }
 
-// Set updates the cache values for the provided key. If the key already exists,
-// its previous values are removed before adding the new values. The operation is thread-safe.
-func (c *InMemoryHashMapCache[K, T, H]) Set(key K, values ...T) {
-	c.vMtx.Lock()
-	defer c.vMtx.Unlock()
-	c.dropKey(key.Keys())
-	c.put(key, values...)
-	c.lastUpdatedKeys[key.String()] = time.Now()
-	c.lastUpdated = time.Now()
-}
-
-// PutQuietly adds values to the cache for the provided key but does so without
+// SetQuietly is an optimized method that adds value to the cache for the provided key but does so without
 // altering the change history. This operation can be used when modifications should not trigger cache change diff.
-func (c *InMemoryHashMapCache[K, T, H]) PutQuietly(key K, values ...T) {
+// This operation is much faster and can be used to optimize cache performance in case you don't want to track changes.
+func (c *InMemoryHashMapCache[K, T, H]) SetQuietly(key K, value T) {
 	c.vMtx.Lock()
 	defer c.vMtx.Unlock()
-	c.addTran(key, values...)
+	c.addTran(key, value)
 	c.lastUpdatedKeys[key.String()] = time.Now()
 	c.lastUpdated = time.Now()
 }
 
-// Get retrieves the values associated with the provided key from the cache.
+// Get retrieves the value associated with the provided key from the cache.
 // The operation is thread-safe and does not alter the change history.
-func (c *InMemoryHashMapCache[K, T, H]) Get(key K) []T {
+func (c *InMemoryHashMapCache[K, T, H]) Get(key K) (*T, bool) {
 	c.vMtx.Lock()
 	defer c.vMtx.Unlock()
 	c.changes = nil
 
-	return c.values[c.toHash(key.Keys())]
-}
-
-// Changes returns a list of keys that have experienced changes in the cache since the last reset.
-func (c *InMemoryHashMapCache[K, T, H]) Changes() []K {
-	return c.changes
+	value, ok := c.values[c.toHash(key.Keys())]
+	return &value, ok
 }
 
 // Drop completely clears the cache, removing all entries. The operation is thread-safe.
@@ -446,7 +137,7 @@ func (c *InMemoryHashMapCache[K, T, H]) Drop() {
 	c.lastUpdatedKeys = make(map[string]time.Time)
 }
 
-// DropKey removes the values associated with the provided key from the cache. The operation is thread-safe.
+// DropKey removes the value associated with the provided key from the cache. The operation is thread-safe.
 func (c *InMemoryHashMapCache[K, T, H]) DropKey(key K) {
 	c.vMtx.Lock()
 	defer c.vMtx.Unlock()
@@ -478,12 +169,12 @@ func (c *InMemoryHashMapCache[K, T, H]) Outdated(key uopt.Opt[K]) bool {
 }
 
 func (c *InMemoryHashMapCache[K, T, H]) dropAll() {
-	c.values = make(map[H][]T)
+	c.values = make(map[H]T)
 	c.changes = nil
 }
 
-func (c *InMemoryHashMapCache[K, T, H]) put(key K, values ...T) {
-	c.addTran(key, values...)
+func (c *InMemoryHashMapCache[K, T, H]) put(key K, value T) {
+	c.addTran(key, value)
 	changes := len(c.changes) == 0
 	found := false
 	for _, diff := range c.changes {
@@ -501,39 +192,14 @@ func (c *InMemoryHashMapCache[K, T, H]) put(key K, values ...T) {
 	}
 }
 
-func (c *InMemoryHashMapCache[K, T, H]) addTran(key K, values ...T) {
+func (c *InMemoryHashMapCache[K, T, H]) addTran(key K, value T) {
 	keys := key.Keys()
-	if len(values) == 0 {
-		return
-	}
 
 	for i := 0; i < len(keys); i++ {
-		hash := c.toHash(keys[:i+1])
-		for _, value := range values {
-			if existing, found := c.values[hash]; found {
-				if ind, entry := uarray.ContainsPredicate[T](existing, func(v *T) bool {
-					return (*v).Equals(value)
-				}); entry == nil {
-					// Collision detected
-					c.values[hash] = append(existing, value)
-				} else {
-					// Else replace, this value is already hashed
-					c.values[hash][ind] = value
-				}
-			} else {
-				c.values[hash] = []T{value}
-			}
-		}
+		c.values[c.toHash(keys[:i+1])] = value
 	}
 }
 
 func (c *InMemoryHashMapCache[K, T, H]) dropKey(keys []int64) {
 	delete(c.values, c.toHash(keys))
-}
-
-func intToBytes(buffer *bytes.Buffer, num int64) []byte {
-	buffer.Reset()
-	_ = binary.Write(buffer, binary.LittleEndian, num)
-
-	return buffer.Bytes()
 }
